+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update shipping data on `orderForm` after shipping simulation.
+- Pre-fill postal code information from `orderForm`'s shipping data.
 
 ## [3.124.0] - 2020-08-27
 ### Added